Plane: Add a feed forward from flaps to pitch
The intent is to approximate the effect of what is typically done on the
RC transmitter. While the integrator will typically catch up to this, it
still is a reactive system rather then proactive, and can have negative
impacts on things like servo auto trim.

This is currently only applied to auto flaps, and autopilot controlled
modes. This could have also been implemented inside of APM_Control.